debugged and finished

+++ hw5/BLK_caching.cpp
@@ -34,7 +34,7 @@ bool BLK_caching::check_hit(unsigned long long adr)
 {
     int hash_set = ((adr & 0x7f8) >> 3);
     int hash_block = adr & 0x7;
-    int adr_used = adr - hash_block;
+    unsigned long long adr_used = adr - hash_block;
     int pos = this->has_adr(adr);
     if (pos != -1)
     {
